feat: use next-auth v5 (v4 doesn't supports next 15)
