+++ cli/src/installers/nextAuth.ts
@@ -6,11 +6,7 @@ import { type AvailableDependencies } from "~/installers/dependencyVersionMap.js
 import { type Installer } from "~/installers/index.js";
 import { addPackageDependency } from "~/utils/addPackageDependency.js";
 
-export const nextAuthInstaller: Installer = ({
-  projectDir,
-  packages,
-  appRouter,
-}) => {
+export const nextAuthInstaller: Installer = ({ projectDir, packages }) => {
   const usingPrisma = packages?.prisma.inUse;
   const usingDrizzle = packages?.drizzle.inUse;
 
@@ -26,17 +22,14 @@ export const nextAuthInstaller: Installer = ({
 
   const extrasDir = path.join(PKG_ROOT, "template/extras");
 
-  const apiHandlerFile = "src/pages/api/auth/[...nextauth].ts";
-  const routeHandlerFile = "src/app/api/auth/[...nextauth]/route.ts";
-  const srcToUse = appRouter ? routeHandlerFile : apiHandlerFile;
+  const apiHandlerFile = "src/app/api/auth/[...nextauth]/route.ts";
 
-  const apiHandlerSrc = path.join(extrasDir, srcToUse);
-  const apiHandlerDest = path.join(projectDir, srcToUse);
+  const apiHandlerSrc = path.join(extrasDir, apiHandlerFile);
+  const apiHandlerDest = path.join(projectDir, apiHandlerFile);
 
   const authConfigSrc = path.join(
     extrasDir,
-    "src/server",
-    appRouter ? "auth-app" : "auth-pages",
+    "src/server/auth",
     usingPrisma
       ? "with-prisma.ts"
       : usingDrizzle
