+++ cli/template/extras/src/app/api/auth/[...nextauth]/route.ts
@@ -1,7 +1,3 @@
-import NextAuth from "next-auth";
+import { handlers } from "~/server/auth";
 
-import { authOptions } from "~/server/auth";
-
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
+export const { GET, POST } = handlers;
