+++ cli/template/extras/src/app/page/with-auth-trpc.tsx
@@ -1,13 +1,13 @@
 import Link from "next/link";
 
 import { LatestPost } from "~/app/_components/post";
-import { getServerAuthSession } from "~/server/auth";
+import { auth } from "~/server/auth";
 import { api, HydrateClient } from "~/trpc/server";
 import styles from "./index.module.css";
 
 export default async function Home() {
   const hello = await api.post.hello({ text: "from tRPC" });
-  const session = await getServerAuthSession();
+  const session = await auth();
 
   if (session?.user) {
     void api.post.getLatest.prefetch();
