--- cli/template/extras/src/pages/api/auth/[...nextauth].ts
@@ -1,5 +1,0 @@
-import NextAuth from "next-auth";
-
-import { authOptions } from "~/server/auth";
-
-export default NextAuth(authOptions);
